Revert "Try stuff"
This reverts commit acb589252622a0cb5afe1aba96ccfc3eed0a90d3.
